requirements.txt

python 3.7+
Flask
tensorflow
Pillow
numpy
scipy
